+++ app/Http/Livewire/Settings/ConfigGroups/Detail.php
@@ -5,6 +5,7 @@ namespace App\Http\Livewire\Settings\ConfigGroups;
 use Livewire\Component;
 use App\Models\ConfigGroup;
 use App\Models\ConfigAppl;
+use App\Models\ConfigUser;
 use Illuminate\Validation\Rule;
 use Lang;
 use Exception;
@@ -17,56 +18,88 @@ class Detail extends Component
     public $action = 'Create';
     public $objectId;
     public $inputs = [];
-    public $applications;
-    public $languages;
     public $status = '';
+    public $applications;
+    public $users;
 
     public function mount($action, $objectId = null)
     {
         $this->action = $action;
         $this->objectId = $objectId;
 
-        $applicationsData = ConfigAppl::GetActiveData();
-
-        $this->applications = $applicationsData->map(function ($data) {
-            return [
-                'label' => $data->code . ' - ' . $data->name,
-                'value' => $data->code,
-            ];
-        })->toArray();
-        $this->inputs['applications'] = $this->applications[0]['value'];
-
+        $this->refreshApplication();
+        $this->refreshUser();
         if (($this->action === 'Edit' || $this->action === 'View') && $this->objectId) {
             $this->object = ConfigGroup::withTrashed()->find($this->objectId);
             $this->status = $this->object->deleted_at ? 'Non-Active' : 'Active';
             $this->VersioNumber = $this->object->version_number;
-            $this->inputs['code'] = $this->object->code;
-            $this->inputs['name']  =  $this->object->name;
+            $this->inputs = populateArrayFromModel($this->object);
         } else {
             $this->object = new ConfigGroup();
         }
     }
+
+    public function refreshApplication()
+    {
+        $applicationsData = ConfigAppl::GetActiveData();
+        if (!$applicationsData->isEmpty()) {
+            $this->applications = $applicationsData->map(function ($data) {
+                return [
+                    'label' => $data->code . ' - ' . $data->name,
+                    'value' => $data->id,
+                ];
+            })->toArray();
+
+            $this->inputs['appl_id'] = $this->applications[0]['value'];
+        } else {
+            $this->applications = [];
+            $this->inputs['appl_id'] = null;
+        }
+    }
+
+    public function refreshUser()
+    {
+        $usersData = ConfigUser::GetActiveData();
+        if (!$usersData->isEmpty()) {
+            $this->users = $usersData->map(function ($data) {
+                return [
+                    'label' => $data->code . ' - ' . $data->name,
+                    'value' => $data->id,
+                ];
+            })->toArray();
+            $this->inputs['user_id'] = '';
+        } else {
+            $this->users = [];
+            $this->inputs['user_id'] = null;
+        }
+    }
+
 
     public function render()
     {
         return view('livewire.settings.config-groups.edit');
     }
 
+    protected $listeners = [
+        'changeStatus'  => 'changeStatus',
+    ];
+
     protected function rules()
     {
         $rules = [
-            'inputs.applications' => 'required|string|min:1|max:50',
+            'inputs.appl_id' => 'required',
+            'inputs.user_id' => 'required',
             'inputs.name' => 'required|string|min:1|max:100',
-            'inputs.code' => [
-                'required',
-                'string',
-                'min:1',
-                'max:50',
-                Rule::unique('config_groups', 'code')
-                    ->ignore($this->object->id)
-                    ->where(function ($query) {
-                    }),
-            ],
+            // 'inputs.code' => [
+            //     'required',
+            //     'string',
+            //     'min:1',
+            //     'max:50',
+            //     Rule::unique('config_groups', 'code')
+            //         ->ignore($this->object->id)
+            //         ->where(function ($query) {
+            //         }),
+            // ],
         ];
         return $rules;
     }
@@ -75,43 +108,61 @@ class Detail extends Component
         'inputs'                => 'Input Group',
         'inputs.*'              => 'Input Group',
         'inputs.code'           => 'Group Code',
-        'inputs.applications'      => 'Group Application Code',
+        'inputs.appl_id'      => 'Application',
+        'inputs.user_id'      => 'User',
         'inputs.name'      => 'Group Name'
     ];
 
     protected function populateObjectArray()
     {
-        return [
-            'code' => $this->inputs['code'],
-            'appl_code' => $this->inputs['applications'],
-            'name' => $this->inputs['name']
-        ];
+        $objectData =  populateModelFromForm($this->object, $this->inputs);
+
+        $application = ConfigAppl::find($this->inputs['appl_id']);
+        $user = ConfigUser::find($this->inputs['user_id']);
+
+        $objectData['appl_code'] = $application->code;
+        $objectData['user_code'] =  $user->code;
+        return $objectData;
+    }
+
+    public function validateForms()
+    {
+        try {
+            $this->validate();
+        } catch (Exception $e) {
+            $this->dispatchBrowserEvent('notify-swal', [
+                'type' => 'error',
+                'message' => Lang::get('generic.error.create', ['object' => $this->object->name, 'message' => $e->getMessage()])
+            ]);
+            throw $e;
+        }
     }
 
     public function Create()
     {
+        $this->validateForms();
         try {
-            $this->validate();
             $objectData = $this->populateObjectArray();
             $this->object = ConfigGroup::create($objectData);
             $this->dispatchBrowserEvent('notify-swal', [
                 'type' => 'success',
                 'message' => Lang::get('generic.success.create', ['object' => $this->inputs['name']])
             ]);
-            $this->inputs = [];
-            $this->inputs['applications'] = $this->applications[0]['value'];
+            $this->reset('inputs');
+            $this->refreshApplication();
+            $this->refreshUser();
         } catch (Exception $e) {
             $this->dispatchBrowserEvent('notify-swal', [
                 'type' => 'error',
-                'message' => Lang::get('generic.error.create', ['object' => "User", 'message' => $e->getMessage()])
+                'message' => Lang::get('generic.error.create', ['object' =>  $this->inputs['name'], 'message' => $e->getMessage()])
             ]);
         }
     }
 
     public function Edit()
     {
+        $this->validateForms();
         try {
-            $this->validate();
 
             if ($this->object) {
                 $this->object->updateObject($this->VersioNumber);
@@ -135,40 +186,32 @@ class Detail extends Component
         }
     }
 
-    public function Disable()
+    public function changeStatus()
     {
         try {
             $this->object->updateObject($this->VersioNumber);
-            $this->object->delete();
-            $this->dispatchBrowserEvent('notify-swal', [
-                'type' => 'success',
-                'message' => Lang::get('generic.success.disable', ['object' => $this->object->name])
-            ]);
-        } catch (Exception $e) {
-            $this->dispatchBrowserEvent('notify-swal', [
-                'type' => 'error',
-                'message' => Lang::get('generic.error.disable', ['object' => $this->object->name, 'message' => $e->getMessage()])
-            ]);
-        }
-        $this->dispatchBrowserEvent('refresh');
-    }
 
-    public function Enable()
-    {
-        try {
-            $this->object->updateObject($this->VersioNumber);
-            $this->object->deleted_at = null;
+            if ($this->object->deleted_at) {
+                $this->object->deleted_at = null;
+                $messageKey = 'generic.success.enable';
+            } else {
+                $this->object->delete();
+                $messageKey = 'generic.success.disable';
+            }
+
             $this->object->save();
+
             $this->dispatchBrowserEvent('notify-swal', [
                 'type' => 'success',
-                'message' => Lang::get('generic.success.enable', ['object' => $this->object->name])
+                'message' => Lang::get($messageKey, ['object' => $this->object->name])
             ]);
         } catch (Exception $e) {
             $this->dispatchBrowserEvent('notify-swal', [
                 'type' => 'error',
-                'message' => Lang::get('generic.error.enable', ['object' => $this->object->name, 'message' => $e->getMessage()])
+                'message' => Lang::get('generic.error.' . ($this->object->deleted_at ? 'enable' : 'disable'), ['object' => $this->object->name, 'message' => $e->getMessage()])
             ]);
         }
+
         $this->dispatchBrowserEvent('refresh');
     }
 }
